Refactoring
* Fix package name issues with Token entity generator strategy
* Moved the event and event.listener packages from com.kossyuzokwe to
com.kossyuzokwe.fantasy
