+++ src/main/java/com/kossyuzokwe/fantasy/event/listener/RegistrationCompleteListener.java
@@ -1,4 +1,4 @@
-package com.kossyuzokwe.event.listener;
+package com.kossyuzokwe.fantasy.event.listener;
 
 import java.util.UUID;
 
@@ -6,12 +6,12 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationListener;
 import org.springframework.mail.SimpleMailMessage;
 
-import com.kossyuzokwe.event.OnResetPasswordEvent;
+import com.kossyuzokwe.fantasy.event.OnRegistrationCompleteEvent;
 import com.kossyuzokwe.fantasy.model.User;
 import com.kossyuzokwe.fantasy.service.EmailService;
 import com.kossyuzokwe.fantasy.service.UserService;
 
-public class ResetPasswordListener implements ApplicationListener<OnResetPasswordEvent> {
+public class RegistrationCompleteListener implements ApplicationListener<OnRegistrationCompleteEvent> {
 	
 	@Autowired
 	private UserService userService;
@@ -20,11 +20,11 @@ public class ResetPasswordListener implements ApplicationListener<OnResetPasswor
 	private EmailService emailService;
 
 	@Override
-	public void onApplicationEvent(OnResetPasswordEvent event) {
+	public void onApplicationEvent(OnRegistrationCompleteEvent event) {
 		User user = event.getUser();
 		String token = UUID.randomUUID().toString();
-		userService.createPasswordResetTokenForUser(user, token);
-		SimpleMailMessage email = emailService.constructResetEmailMessage(event, user, token);
+		userService.createVerificationTokenForUser(user, token);
+		SimpleMailMessage email = emailService.constructVerifyEmailMessage(event, user, token);
 		emailService.send(email);
 	}
