+++ src/main/java/com/kossyuzokwe/fantasy/event/listener/ResetPasswordListener.java
@@ -1,4 +1,4 @@
-package com.kossyuzokwe.event.listener;
+package com.kossyuzokwe.fantasy.event.listener;
 
 import java.util.UUID;
 
@@ -6,12 +6,12 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationListener;
 import org.springframework.mail.SimpleMailMessage;
 
-import com.kossyuzokwe.event.OnRegistrationCompleteEvent;
+import com.kossyuzokwe.fantasy.event.OnResetPasswordEvent;
 import com.kossyuzokwe.fantasy.model.User;
 import com.kossyuzokwe.fantasy.service.EmailService;
 import com.kossyuzokwe.fantasy.service.UserService;
 
-public class RegistrationCompleteListener implements ApplicationListener<OnRegistrationCompleteEvent> {
+public class ResetPasswordListener implements ApplicationListener<OnResetPasswordEvent> {
 	
 	@Autowired
 	private UserService userService;
@@ -20,11 +20,11 @@ public class RegistrationCompleteListener implements ApplicationListener<OnRegis
 	private EmailService emailService;
 
 	@Override
-	public void onApplicationEvent(OnRegistrationCompleteEvent event) {
+	public void onApplicationEvent(OnResetPasswordEvent event) {
 		User user = event.getUser();
 		String token = UUID.randomUUID().toString();
-		userService.createVerificationTokenForUser(user, token);
-		SimpleMailMessage email = emailService.constructVerifyEmailMessage(event, user, token);
+		userService.createPasswordResetTokenForUser(user, token);
+		SimpleMailMessage email = emailService.constructResetEmailMessage(event, user, token);
 		emailService.send(email);
 	}
