+++ src/main/java/com/kossyuzokwe/fantasy/model/VerificationToken.java
@@ -23,7 +23,7 @@ public class VerificationToken {
 	private static final int EXPIRATION = 60 * 24;
 
 	@Id
-	@GenericGenerator(name = "sequence_object_id", strategy = "com.kossyuzokwe.util.ObjectIdGenerator")
+	@GenericGenerator(name = "sequence_object_id", strategy = "com.kossyuzokwe.fantasy.util.ObjectIdGenerator")
 	@GeneratedValue(generator = "sequence_object_id")
 	@Column(name = "verification_id")
 	private String verificationId;
